requirements.txt
flask
opencv-contrib-python
numpy
pillow
